Move the core implementation to Rust to improve realtime performance. (#11)
* update: port actuator backends to Rust with Python wrappers

* fix: disable local loopback

* fix: example

* fix: backend error handling and checked command encoding

* update: mit interface

* add: docstrings

* Update actuator_control/api.py


---------

+++ examples/actuator_characterization/generate_test_signal.py
@@ -264,7 +264,7 @@ class CharacterizationTest:
 
         signal = np.concatenate(blocks)
 
-        times = np.arange(signal.shape[0], dtype=np.float64) * dt
+        times = np.arange(signal.shape[0], dtype=np.float32) * dt
         self.times = times.astype(np.float32)
         self.signal = signal.astype(np.float32)
